Machine Learning first project
RESSEL (Reliable Semi-Supervised Ensemble Learning) is a learning method that combines semi-supervised learning with ensemble learning.

In this approach, both labeled and unlabeled data are used to train the model. The goal is to leverage unlabeled data to improve supervised learning. The key idea behind RESSEL is that by utilizing a set of unlabeled samples, the classification process can be enhanced, leading to better prediction accuracy.